fix streaming update perf (#519)
In streaming xact, when processing deletion,
we need to do a index lookup in foreground on everything in current transaction, both in-memory and flushed rows (since they are not visible to snapshot thread yet).
Index lookup is async function that may involve file IO, so it is relatively costly.
But it is very unlikely a row in current transaction is deleted, so most of the check is indeed wasted.

Fix it by adding a bloomfliter.