fix multi column key (#566)
We were calling equals_record_batch_at_offset with wrong input
(the function is expecting full row and caller calls with extracted values).

Fix the mismatch.
Also fix a small perf inefficiency (we don't need to read record batch when the key is stored in memory)

Modified test config so we will hit the case.