fix another edge case in deletion index lookup (#1464)
Deletion of the same value need to be process in the LSN order.
Otherwise, the deletion with larger LSN will see multiple candidates, and the one with smaller LSN will see less.

This normally never happens, because in general we add records to new_deletions in LSN order.

But streamingXact that delete's row in memslice,
and the memslice got flushed before streamingXact finishes,
may cause them to be added in different order.